+++ src/App/Bundle/AppBundle/DependencyInjection/AppAppExtension.php
@@ -24,5 +24,6 @@ class AppAppExtension extends Extension
 
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yml');
+        $loader->load('FAQ/services.yml');
     }
 }