+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 Features:
 
+* Add a FAQ page containing many questions
 * Create an empty changelog (and a paradox)